+++ src/index.ts
@@ -3,6 +3,7 @@ import axios, { AxiosInstance } from "axios";
 import Apps from "./Apps";
 import Alerts from "./Alerts";
 import Addons from "./Addons";
+import AuditLogs from "./AuditLogs";
 import Autoscalers from "./Autoscalers";
 import Billing from "./Billing";
 import Collaborators from "./Collaborators";
@@ -81,6 +82,7 @@ export class Client {
   Addons = new Addons(this);
   Alerts = new Alerts(this);
   Autoscalers = new Autoscalers(this);
+  AuditLogs = new AuditLogs(this);
   Billing = new Billing(this);
   Collaborators = new Collaborators(this);
   Containers = new Containers(this);